pipecat-server/requirements.txt
# Pipecat Voice Coach Server Requirements

# Core Pipecat framework
pipecat-ai

# WebSocket support
websockets

# Gemini Live AI Service
google-generativeai
python-dotenv

# Audio processing (if needed)
# Note: Gemini Live handles audio directly

# Web framework for additional endpoints (optional)
fastapi
uvicorn

# Logging and utilities
python-json-logger
